Change groups service to use service traceprovider
This changes the groups service to use the service traceprovider.
